Remove explicit cacheray interaction from testbench
The conditional use of cacheray was always enabled, so remove the
conditionality entirely. It can be implemented in the build system by
adding -fsanitize=thread and -lcacheray conditionally instead.

I have not been able to test this change because the testbench seems to
be hard-coded to hastrom's machine. Oh well.