Remove cacheray runtime API
The runtime self-initializes, and there's no need for target programs to
call into it.

This also removes the manual buffering and callback machinery from the
runtime, and relies entirely on libc file buffering.

Longer-term it would be nice to have an application API for RTTA events,
something like ASAN's __asan_poison_memory_region, but such an API
should probably use double-underscore reserved names, so wait until
we've settled on a convention for that.